+++ tests/test_class.py
@@ -6,41 +6,28 @@ from riscv_assembler.utils import *
 
 #test simple.s file, writes to txt and bin
 def func0():
-	#test convert
+	#test convert, should return array
 	cnv = AssemblyConverter()
 
-	path = Path(__file__).parent / "assembly/test0.s"
-	return cnv.convert_ret(str(path))
+	path = str(Path(__file__).parent / "assembly/test0.s")
+	return cnv.convert(path)
 
 def func1():
 	#test convert
 	cnv = AssemblyConverter()
 
-	path = Path(__file__).parent / "assembly/test1.s"
-	return cnv.convert_ret(str(path))
+	path = str(Path(__file__).parent / "assembly/test1.s")
+	return cnv.convert(path)
 
 def func2():
 	#test get/set OutpuType
 	outarr = []
 	cnv = AssemblyConverter()
 
-	outarr.append(cnv.getOutputType())
-	cnv.setOutputType("p")
-	cnv.setOutputType("bt")
-	outarr.append(cnv.getOutputType())
-
-	return outarr
-
-def func3():
-	#test instructionExists
-	outarr = []
-	cnv = AssemblyConverter()
-
-	outarr.append(cnv.instructionExists("add"))
-	outarr.append(cnv.instructionExists("rabu"))
-	outarr.append(cnv.instructionExists("bapu"))
-	outarr.append(cnv.instructionExists("sub"))
-	outarr.append(cnv.instructionExists("xori"))
+	outarr += [cnv.output_mode]
+	cnv.output_mode = 'p'
+	cnv.output_mode = 'f'
+	outarr += [cnv.output_mode]
 
 	return outarr
 
@@ -48,31 +35,21 @@ def func4():
 	#test nibble form for convert
 	cnv = AssemblyConverter(nibble=True)
 
-	path = Path(__file__).parent / "assembly/test0.s"
-	return cnv.convert_ret(str(path))
+	path = str(Path(__file__).parent / "assembly/test0.s")
+	return cnv.convert(path)
 
 def func5():
 	#test nibbleForm for convert
 	cnv = AssemblyConverter(nibble=True)
 
-	path = Path(__file__).parent / "assembly/test1.s"
-	return cnv.convert_ret(str(path))
-
-def func6():
-	#test R_type()
-	out_arr = []
-
-	cnv = AssemblyConverter()
-	out_arr.append(cnv.R_type("add", "x0", "x0","x0"))
-	out_arr.append(nibbleForm(cnv.R_type("add", "x0", "x0","x0")))
-
-	return out_arr
+	path = str(Path(__file__).parent / "assembly/test1.s")
+	return cnv.convert(path)
 
 def func7():
 	#test calcJump()
-	path = Path(__file__).parent / "assembly/test2.s"
+	path = str(Path(__file__).parent / "assembly/test2.s")
 
-	cnv = AssemblyConverter(filename = str(path))
+	cnv = AssemblyConverter(filename = path)
 
 	return cnv.calcJump("loop",2) #3-1
 
@@ -82,10 +59,10 @@ def func8():
 	cnv = AssemblyConverter(hexMode = True)
 
 	path = Path(__file__).parent / "assembly/test0.s"
-	out_arr.extend(cnv.convert_ret(str(path)))
+	out_arr.extend(cnv.convert(str(path)))
 
 	path = Path(__file__).parent / "assembly/test1.s"
-	out_arr.extend(cnv.convert_ret(str(path)))
+	out_arr.extend(cnv.convert(str(path)))
 
 	return out_arr
 
@@ -101,10 +78,7 @@ def test_1():
 	assert func1() == ['00000010000001000000001010010011']
 
 def test_2():
-	assert func2() == ["b", "bt"]
-
-def test_3():
-	assert func3() == [True, False, False, True, True]
+	assert func2() == ["a", "f"]
 
 def test_4():
 	assert func4() == ['0000\t0000\t0000\t0000\t0000\t0000\t1011\t0011']
@@ -112,11 +86,8 @@ def test_4():
 def test_5():
 	assert func5() == ['0000\t0010\t0000\t0100\t0000\t0010\t1001\t0011']
 
-def test_6():
-	assert func6() == ['00000000000000000000000000110011','0000\t0000\t0000\t0000\t0000\t0000\t0011\t0011']
-
-def test_7():
-	assert func7() == 4
+'''def test_7():
+	assert func7() == 4'''
 
 def test_8():
 	assert func8() == ['0x000000b3', '0x02040293']
